requirements.txt
# EnderTrack Minimal v1 - Python dependencies
#
# Toutes les dépendances sont incluses dans vendor/ (zéro install).
# Ce fichier est fourni uniquement si vous préférez installer via pip :
#
#   pip install -r requirements.txt

flask>=2.0
flask-cors>=3.0
pyserial>=3.5
